Allow message types of up to 8 characters

+++ artio-codecs/src/main/java/uk/co/real_logic/artio/builder/Encoder.java
@@ -53,7 +53,7 @@ public interface Encoder
      */
     void reset();
 
-    int messageType();
+    long messageType();
 
     SessionHeaderEncoder header();
