+++ artio-system-tests/src/perf/java/uk/co/real_logic/artio/OtfInlineableAcceptor.java
@@ -54,7 +54,7 @@ public final class OtfInlineableAcceptor implements OtfMessageAcceptor
 
     public boolean onError(
         final ValidationError error,
-        final int messageType,
+        final long messageType,
         final int tagNumber,
         final AsciiFieldFlyweight value)
     {
